requirements.txt
aiohttp
openai
pillow
python-dotenv
ttkbootstrap
